Rename matching router to match and check for user's last date
